combine pr info & add diff url enable (#9)
* combine pr info & add diff url enable

* add diff info

---------
